requirements.txt
qdrant-client>=1.11.0
sentence-transformers
transformers>=4.40
torch
accelerate
numpy
python-dotenv